.history/backend/requirements_20260103003754.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
celery==5.3.6
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0
python-dotenv==1.0.0
tenacity==8.2.3
